impl aggregattype try_from
