+++ pg_search/src/postgres/customscan/aggregatescan/mod.rs
@@ -634,7 +634,7 @@ fn extract_aggregates(args: &CreateUpperPathsHookArgs) -> Option<Vec<AggregateTy
                     aggregate_types.push(AggregateType::Count { missing: None });
                 } else {
                     // Check for other aggregate functions with arguments
-                    let agg_type = identify_aggregate_function(aggref, relation_oid)?;
+                    let agg_type = AggregateType::try_from(aggref, relation_oid)?;
                     aggregate_types.push(agg_type);
                 }
             } else {
@@ -650,99 +650,6 @@ fn extract_aggregates(args: &CreateUpperPathsHookArgs) -> Option<Vec<AggregateTy
     // an empty vector instead of rejecting the plan.
 
     Some(aggregate_types)
-}
-
-/// Identify an aggregate function by its OID and extract field name from its arguments
-unsafe fn identify_aggregate_function(
-    aggref: *mut pg_sys::Aggref,
-    relation_oid: pg_sys::Oid,
-) -> Option<AggregateType> {
-    let aggfnoid = (*aggref).aggfnoid;
-
-    // Get the function name to identify the aggregate
-    let func_name = get_aggregate_function_name(aggfnoid)?;
-
-    // Extract the field name from the first argument
-    let field_name = extract_field_name_from_aggref(aggref, relation_oid);
-
-    match func_name {
-        "count" => Some(AggregateType::Count { missing: None }),
-        "sum" => Some(AggregateType::Sum { field: field_name?, missing: None }),
-        "avg" => Some(AggregateType::Avg { field: field_name?, missing: None }),
-        "min" => Some(AggregateType::Min { field: field_name?, missing: None }),
-        "max" => Some(AggregateType::Max { field: field_name?, missing: None }),
-        _ => {
-            pgrx::debug1!("Unsupported aggregate function: {func_name}");
-            None
-        }
-    }
-}
-
-/// Get the name of an aggregate function from its OID
-unsafe fn get_aggregate_function_name(aggfnoid: pg_sys::Oid) -> Option<&'static str> {
-    use pgrx::pg_sys::{
-        F_AVG_FLOAT4, F_AVG_FLOAT8, F_AVG_INT2, F_AVG_INT4, F_AVG_INT8, F_AVG_NUMERIC, F_COUNT_ANY,
-        F_MAX_DATE, F_MAX_FLOAT4, F_MAX_FLOAT8, F_MAX_INT2, F_MAX_INT4, F_MAX_INT8, F_MAX_NUMERIC,
-        F_MAX_TIME, F_MAX_TIMESTAMP, F_MAX_TIMESTAMPTZ, F_MAX_TIMETZ, F_MIN_DATE, F_MIN_FLOAT4,
-        F_MIN_FLOAT8, F_MIN_INT2, F_MIN_INT4, F_MIN_INT8, F_MIN_MONEY, F_MIN_NUMERIC, F_MIN_TIME,
-        F_MIN_TIMESTAMP, F_MIN_TIMESTAMPTZ, F_MIN_TIMETZ, F_SUM_FLOAT4, F_SUM_FLOAT8, F_SUM_INT2,
-        F_SUM_INT4, F_SUM_INT8, F_SUM_NUMERIC,
-    };
-    // Use well-known PostgreSQL function OIDs for standard aggregates
-    // These are consistent across PostgreSQL versions
-    match aggfnoid.to_u32() {
-        F_AVG_INT8 | F_AVG_INT4 | F_AVG_INT2 | F_AVG_NUMERIC | F_AVG_FLOAT4 | F_AVG_FLOAT8 => {
-            Some("avg")
-        }
-        F_SUM_INT8 | F_SUM_INT4 | F_SUM_INT2 | F_SUM_FLOAT4 | F_SUM_FLOAT8 | F_SUM_NUMERIC => {
-            Some("sum")
-        }
-        F_MAX_INT8 | F_MAX_INT4 | F_MAX_INT2 | F_MAX_FLOAT4 | F_MAX_FLOAT8 | F_MAX_DATE
-        | F_MAX_TIME | F_MAX_TIMETZ | F_MAX_TIMESTAMP | F_MAX_TIMESTAMPTZ | F_MAX_NUMERIC => {
-            Some("max")
-        }
-        F_MIN_INT8 | F_MIN_INT4 | F_MIN_INT2 | F_MIN_FLOAT4 | F_MIN_FLOAT8 | F_MIN_DATE
-        | F_MIN_TIME | F_MIN_TIMETZ | F_MIN_MONEY | F_MIN_TIMESTAMP | F_MIN_TIMESTAMPTZ
-        | F_MIN_NUMERIC => Some("min"),
-        F_COUNT_ANY => Some("count"),
-        _ => {
-            // For unknown function OIDs, we'll reject them for now
-            pgrx::debug1!("Unknown aggregate function OID: {}", aggfnoid.to_u32());
-            None
-        }
-    }
-}
-
-/// Extract field name from the first argument of an aggregate function
-unsafe fn extract_field_name_from_aggref(
-    aggref: *mut pg_sys::Aggref,
-    relation_oid: pg_sys::Oid,
-) -> Option<String> {
-    let args = PgList::<pg_sys::TargetEntry>::from_pg((*aggref).args);
-    if args.is_empty() {
-        return None;
-    }
-
-    let first_arg = args.get_ptr(0)?;
-    if let Some(var) = nodecast!(Var, T_Var, (*first_arg).expr) {
-        return get_var_field_name(var, relation_oid);
-    }
-
-    None
-}
-
-/// Get the field name from a Var node
-unsafe fn get_var_field_name(var: *mut pg_sys::Var, relation_oid: pg_sys::Oid) -> Option<String> {
-    let varattno = (*var).varattno;
-
-    // Get the actual column name from the relation
-    let attname = pg_sys::get_attname(relation_oid, varattno, false);
-    if !attname.is_null() {
-        let name = std::ffi::CStr::from_ptr(attname).to_str().ok()?;
-        return Some(name.to_string());
-    }
-
-    None
 }
 
 /// Replace any T_Aggref expressions in the target list with T_FuncExpr placeholders
